Add files via upload
According to WHO statistics, liver cancer is the second leading cause of mortality among all cancers. A computed tomography scan is usually used to diagnose it. Deep learning algorithms are now being utilized to segment the liver and its tumour using computed tomography (CT) scan pictures. 

For many clinical applications in the medical sciences, segmenting the abdominal CT scan image is a crucial task. Automatic solutions to this problem are crucial because they save time, and effort, give professionals quality support and enable early tumour detection and identification. For the application of segmenting the liver region and tumour area to be useful, it must be extremely accurate, precise, and confirmed using metrics used in medical image processing, so that a radiologist can complete segmentation rapidly and independently. Hence, we are able to build this we an accuracy of 99.9%.